server/pyproject.toml
[tool.poetry]
package-mode = false

[project]
name = "dev-archive"
version = "0.1.0"
description = "RAG pipeline to query and analyze software engineering case studies"
authors = [
    {name = "Danny C",email = " 35072063+dchen705@users.noreply.github.com"}
]
requires-python = ">=3.13,<4.0"
dependencies = [
    "python-dotenv (>=1.2.2,<2.0.0)",
    "openai (>=2.30.0,<3.0.0)",
    "pinecone (>=8.1.0,<9.0.0)",
    "tiktoken (>=0.12.0,<0.13.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
